+++ Doctrine/Phpcr/SimpleBlock.php
@@ -2,10 +2,12 @@
 
 namespace Symfony\Cmf\Bundle\BlockBundle\Doctrine\Phpcr;
 
+use Symfony\Cmf\Bundle\CoreBundle\Translatable\TranslatableInterface;
+
 /**
  * Block that contains hypertext and a title
  */
-class SimpleBlock extends AbstractBlock
+class SimpleBlock extends AbstractBlock implements TranslatableInterface
 {
     /**
      * @var string